Stop migrating .wskprops to credentials
This removes the one remaining place where we were trusting .wskprops.
It was happening only when a user had .wskprops but no credential
store.  We assumed the user was already a Nimbella user, but, in fact,
we were not checking for that.  The 'migration' has outlived its
usefulness anyway.  Issue #76