Lock down octokit dependency to head off breakage
There was a change in the organization of the octokit API that caused
nimbella-cli builds to break.  I should probably update to the latest
version because the usage of this API will increase.   But, for the
moment using a closed dependency in place of an open one will fix the
problem.