Protect against mis-use of nim auth login --auth
If the login --auth matches a namespace already in the cred store, it
is disallowed other than the special case of being invoked from
`nimadmin` using a hidden flag.  The customer is advised to logout
first or login without --auth.  This guards against accidentally
destroying storage credentials by what might appear to be a harmlessly
redundant action.  Issue #148