Two needed changes to deploy tests 'runAutomated'
1.  The ordering is changed to put 'testMainActions' last.  It is the
most likely to fail for reasons unrelated to the deployer and so its
failure should not prevent the demos test from running.  2.  The step
to logout of one of the namespaces, needed for one of the tests, is now
unconditional and not just when the --ongcp flag is used.  That's
because the test namespace is actually on nimgcp regardless of where
the test is otherwise running.  In Jenkins, the logout will cause a
small error message near the start of the run but not stop the run
(I hope).