Refactor nimbella-cli into a deployer library package and a command package (#283)
* Separate nimbella-deployer as its own package

Issue #282

* Move NimBaseCommand into the deployer component

Issue #282

* Builds for the new nim internal structure

Not quite working yet.  See issue #282

* Identify the root package to oclif when invoking

This is now necessary because NimBaseCommand is contributed by a
dependent package.  Issue #282.

* Default source for deployer now preview, not test

Issue #282

* Revise build.sh to cover the new scenarios

Not quite right yet for linking to the main build.  Issue #282

* Final (?) changes to build.sh

Issue #282

* Export nimbellaDir (used by commander-cli)

Issue #282

* Re-enable commander, temp use branch version

The version of commander being incorporated is on the 'use-library'
branch of commander-cli.  This is for sanity checking.  That branch
will be a PR and, once merged, this branch can probably also be merged.
Issue #282

* sensitiveNamespaces now part of deployer only

* Fix directory nesting bug from deployer refactor

* Move runNimCommand back to main library

* Re-export deployer library in legacy location

* Adjust API testing and new test for command API

* Switch to commander master commit

* Update API doc to reflect refactoring of libraries

* Default build should not depend on remote tarballs

PR #283

* Ensure uploaded preview tarballs are not cached

* Add dependencies to deployer, make self-contained

Before it was only building by virtue of the fact that usually there
was a populated node_modules in the parent.  In a fresh clone this is
not true.  Having duplicate dependencies set off some problem in 'tsc'
that caused it to demand some extra type marks.  I'm not sure why
exactly but the new marks are mostly accurate and do no harm.

* Aio dependency is now in two package.jsons

* Changes needed to support workbench build

Issue #282

* Might as well use the latest commander when merged