Remote building (incorporate public repo changes) (#317)
* Support added dependencies for remote build

Issue #163

* Minor fix to API doc

(on branch, will get merged with the branch)

* Add a local test to mock remote slice deployment

* WIP for remote building.  Interim commit.

* Get remote building working for go runtime

Issue #163

* Remote go build test now invokes /bin/compile

Issue #163

* Add test for swift remote build

* WIP for adding Java to remote building

* Temp tweak to go build (not yet working)

We want to find a uniform default build that works for both go and
swift.

* Trivial test cases for "bad build" error handling

* Go back to previous.  Change buys us nothing.

This reverts commit 3dbcdddfe6d3483d8a5647dbf4798d49c13e8128.

* Simplify go and swift illustrative builds

They seem to differ necessarily.   Which is ok because these are the
only two runtimes in which we are likely to install this logic
permanently.  But, the swift one needs more work to handle multi-file.

* Make commander hash maintenance the same as aio's

* Change repo interlock of build to be more flexible

* Interlock repos (now we can do that on branches)